+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - font-family and font-size to charts [167](https://github.com/greenbone/pheme/pull/167)
 ### Changed
 - replaced hardcoded 175 in favor of max len hostname * font size * 1.25 in bar chart [167](https://github.com/greenbone/pheme/pull/167)
+- legend on the middle left instead of middle bottom [170](https://github.com/greenbone/pheme/pull/170)
 ### Deprecated
 ### Removed
 ### Fixed